+++ src/main/java/use_cases/DataAccess.java
@@ -1,7 +1,6 @@
 package use_cases;
 
 import entities.HumanPlayer;
-import usecases.PlayerInformation;
 
 import java.io.IOException;
