+++ src/test/java/database/SaveLoadPlayerTest.java
@@ -4,7 +4,7 @@ import entities.*;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import usecases.PlayerInformation;
+import use_cases.PlayerInformation;
 
 import java.io.*;
 import java.util.*;
